Add "implicit has" option for optional messages
For some applications, it is helpful if messages are optional
when decoding the wire format, but assumed to always be present
when encoding.  This reduces the size of structures required by
removing the "has_XXXX" fields, as well as the code size and
effort by removing maintenance of the "has_XXXX" fields.