cmake: Clean up the install hierarchy and install the generator
This is a fairly extensive change that removes some of the unecessary
vendoring that is done and changes the namespaces for the nanopb options
proto. It cleans up the cmake build so that nanopb can be installed into
a linux userspace, and have programs depend on it.

This also breaks out the configuration options from pb.h into a separate
config.h which is not provided by default. If you use the cmake build
system this will be automatically provided with the most compatible
settings.

Signed-off-by: William A. Kennington III <wak@google.com>