Handle a single non-packed repeated fixed count field
Allow the elements of that single non-packed repeated fixed count field
to be interleaved among other (non-packed repeated fixed count) fields.
Multiple non-packed repeated fixed count fields are allowed if the
elements are not interleaved.

An error is returned if the count of a repeated fixed count field is not
equal the expected count.