Fix CMake options file dependency
Restore dependency of generated nanopb source and header files on an
options file, if it exists in the same directory and has the name
filename without extension.

Add nanopb CMake variable NANOPB_DEPENDS which is a list of files to be
manually set as dependencies for the generated source and header files.
If these dependencies are options files, the corresponding directory is
passed as an options path.