Fix compiler errors with complex oneof hierarchy and sizeof() (#610)
Previously the generator would refer to internal symbols generated
by OneOf.encoded_size(). However the include file might not have
those symbols if they were resolved at generator time.

This commit changes it so that the generator only refers to symbols
that are part of public API for the other include files.