Improve optimization for little-endian platforms.
Previously there was a fast path for little endian platforms
in pb_decode_fixed64() but not in pb_encode_fixed64(). Also
the macros used for the check didn't trigger on GCC.

Macro checks were expanded to cover all common compilers and
now it is possible to specify PB_LITTLE_ENDIAN_8BIT manually
if it is not automatically detected.