Rename pb_i/ostream_t to pb_en/decode_ctx_t
To permit e.g. allocator state to be part of the structure, streams
are being replaced with encode & decode context structures. As part
of this the structures are renamed, which this commit does.

This commit also reworks substream handling so that the context
variable is passed unchanged to callbacks, unlike how streams were
previously copied to substreams.