Add pb_walk() function for message nesting
Instead of relying on C recursion, pb_walk() manages a custom state stack
that stores minimal amount of data for each level. This reduces stack
usage for nested messages.

In addition it will make it possible to compile nanopb without any C recursion.

This commit adds the basic API and uses it for pb_encode().