Add struct size check to pb_en/decode()
This can detect the common mistake of passing wrong _fields array
to the function. It also helps with different compilation options
between files, which could cause struct layout to change.
Because the check is based on sizeof(), it may not catch every
mistake.

Also reorganized how pb_en/decode_ex() flags are handled.
They'll be part of the context now.

Also preparing for using context allocator in pb_release().